+++ R/wf_delete.R
@@ -6,6 +6,8 @@
 #' used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}
 #' @param url url to query
 #' @param verbose show feedback on processing
+#' @param type character, one of \code{ecmwf} or \code{cds} depending
+#' on the data set to be deleted.
 #' @keywords data download, climate, re-analysis
 #' @seealso \code{\link[ecmwfr]{wf_set_key}}
 #' \code{\link[ecmwfr]{wf_transfer}}
@@ -20,10 +22,12 @@
 #' # get key
 #' wf_get_key(email = "test@mail.com")
 #'}
+# TODO: example might need an update (if even required).
 
 wf_delete <- function(
   email,
   url,
+  type = "ecmwf",
   verbose = TRUE
 ){
 
@@ -32,18 +36,38 @@ wf_delete <- function(
     stop("Please provide ECMWF login email / url!")
   }
 
-  # get key from email
-  key <- wf_get_key(email)
+  # Checking input argument 'type'
+  type <- match.arg(type, c("ecmwf", "cds"))
+
+  # get key
+  if(is.null(email)) {
+    tmp   <- get(sprintf("%s_key_from_file", type))(verbose)
+    email <- tmp$email; key <- tmp$key; rm(tmp)
+  } else {
+    key <- get(sprintf("%s_get_key", type))(email)
+  }
 
   # remove a queued download
-  response <- httr::DELETE(
-    url,
-    httr::add_headers(
-      "Accept" = "application/json",
-      "Content-Type" = "application/json",
-      "From" = email,
-      "X-ECMWF-KEY" = key)
-  )
+  # Differs for ecmwf and cds requests.
+  # For CDS: note that 'email' is simply a copy of 'user'
+  if(type == "cds") {
+    response <- httr::DELETE(
+      url,
+      httr::authenticate(email, key),
+      httr::add_headers(
+        "Accept" = "application/json",
+        "Content-Type" = "application/json")
+    )
+  } else {
+    response <- httr::DELETE(
+      url,
+      httr::add_headers(
+        "Accept" = "application/json",
+        "Content-Type" = "application/json",
+        "From" = email,
+        "X-ECMWF-KEY" = key)
+    )
+  }
 
   # check purging of request
   if(response$status == 204){